+++ plugins/src/test/scala/com/stratio/sparta/plugin/workflow/transformation/dropDuplicates/DropDuplicatesTransformStepStreamingIT.scala
@@ -1,0 +1,144 @@
+/*
+ * © 2017 Stratio Big Data Inc., Sucursal en España. All rights reserved.
+ *
+ * This software – including all its source code – contains proprietary information of Stratio Big Data Inc., Sucursal en España and may not be revealed, sold, transferred, modified, distributed or otherwise made available, licensed or sublicensed to third parties; nor reverse engineered, disassembled or decompiled, without express written authorization from Stratio Big Data Inc., Sucursal en España.
+ */
+package com.stratio.sparta.plugin.workflow.transformation.dropDuplicates
+
+import com.stratio.sparta.plugin.TemporalSparkContext
+import com.stratio.sparta.sdk.DistributedMonad.DistributedMonadImplicits
+import com.stratio.sparta.sdk.workflow.enumerators.SaveModeEnum
+import com.stratio.sparta.sdk.workflow.step.{OutputOptions, TransformationStepManagement}
+import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.Row
+import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema
+import org.apache.spark.sql.types.{DoubleType, StringType, StructField, StructType}
+import org.junit.runner.RunWith
+import org.scalatest.Matchers
+import org.scalatest.junit.JUnitRunner
+
+import scala.collection.mutable
+
+@RunWith(classOf[JUnitRunner])
+class DropDuplicatesTransformStepStreamingIT extends TemporalSparkContext with Matchers with DistributedMonadImplicits {
+
+  "A DropDuplicatesTransformStepStreaming" should "don't remove duplicate events from input Stream" in {
+    val schema = StructType(Seq(StructField("color", StringType), StructField("price", DoubleType)))
+    val dataQueue1 = new mutable.Queue[RDD[Row]]()
+    val data1 = Seq(
+      new GenericRowWithSchema(Array("blue", 12.1), schema),
+      new GenericRowWithSchema(Array("red", 12.2), schema),
+      new GenericRowWithSchema(Array("red", 12.3), schema)
+    )
+    dataQueue1 += sc.parallelize(data1)
+    val stream1 = ssc.queueStream(dataQueue1)
+    val inputData = Map("step1" -> stream1)
+    val outputOptions = OutputOptions(SaveModeEnum.Append, "stepName", "tableName", None, None)
+    val result = new DropDuplicatesTransformStepStreaming(
+      "dummy",
+      outputOptions,
+      TransformationStepManagement(),
+      Option(ssc),
+      sparkSession,
+      Map()
+    ).transform(inputData)
+    val totalEvents = ssc.sparkContext.accumulator(0L, "Number of events received")
+
+    result.ds.foreachRDD(rdd => {
+      val streamingEvents = rdd.count()
+      log.info(s" EVENTS COUNT : \t $streamingEvents")
+      totalEvents += streamingEvents
+      log.info(s" TOTAL EVENTS : \t $totalEvents")
+      val streamingRegisters = rdd.collect()
+      if (!rdd.isEmpty())
+        streamingRegisters.foreach(row => assert(data1.contains(row)))
+    })
+    ssc.start()
+    ssc.awaitTerminationOrTimeout(timeoutStreaming)
+    ssc.stop()
+
+    assert(totalEvents.value === 3)
+  }
+
+  "A DropDuplicatesTransformStepStreaming" should "remove duplicate events from input Stream" in {
+    val schema = StructType(Seq(StructField("color", StringType), StructField("price", DoubleType)))
+    val dataQueue1 = new mutable.Queue[RDD[Row]]()
+    val data1 = Seq(
+      new GenericRowWithSchema(Array("blue", 12.1), schema),
+      new GenericRowWithSchema(Array("red", 12.2), schema),
+      new GenericRowWithSchema(Array("red", 12.2), schema)
+    )
+    dataQueue1 += sc.parallelize(data1)
+    val stream1 = ssc.queueStream(dataQueue1)
+    val inputData = Map("step1" -> stream1)
+    val outputOptions = OutputOptions(SaveModeEnum.Append, "stepName", "tableName", None, None)
+    val result = new DropDuplicatesTransformStepStreaming(
+      "dummy",
+      outputOptions,
+      TransformationStepManagement(),
+      Option(ssc),
+      sparkSession,
+      Map()
+    ).transform(inputData)
+    val totalEvents = ssc.sparkContext.accumulator(0L, "Number of events received")
+
+    result.ds.foreachRDD(rdd => {
+      val streamingEvents = rdd.count()
+      log.info(s" EVENTS COUNT : \t $streamingEvents")
+      totalEvents += streamingEvents
+      log.info(s" TOTAL EVENTS : \t $totalEvents")
+      val streamingRegisters = rdd.collect()
+      if (!rdd.isEmpty())
+        streamingRegisters.foreach(row => assert(data1.contains(row)))
+    })
+    ssc.start()
+    ssc.awaitTerminationOrTimeout(timeoutStreaming)
+    ssc.stop()
+
+    assert(totalEvents.value === 2)
+  }
+
+  "A DropDuplicatesTransformStepStreaming" should "remove duplicate events from input Batch based on columns" in {
+    val schema = StructType(Seq(StructField("color", StringType), StructField("price", DoubleType)))
+    val dataQueue1 = new mutable.Queue[RDD[Row]]()
+    val data1 = Seq(
+      new GenericRowWithSchema(Array("blue", 12.1), schema),
+      new GenericRowWithSchema(Array("red", 12.2), schema),
+      new GenericRowWithSchema(Array("red", 12.3), schema)
+    )
+    dataQueue1 += sc.parallelize(data1)
+    val stream1 = ssc.queueStream(dataQueue1)
+    val inputData = Map("step1" -> stream1)
+    val outputOptions = OutputOptions(SaveModeEnum.Append, "stepName", "tableName", None, None)
+    val columns =
+      s"""[
+         |{
+         |   "name":"color"
+         |}]
+         | """.stripMargin
+    val result = new DropDuplicatesTransformStepStreaming(
+      "dummy",
+      outputOptions,
+      TransformationStepManagement(),
+      Option(ssc),
+      sparkSession,
+      Map("columns" -> columns)
+    ).transform(inputData)
+    val totalEvents = ssc.sparkContext.accumulator(0L, "Number of events received")
+
+    result.ds.foreachRDD(rdd => {
+      val streamingEvents = rdd.count()
+      log.info(s" EVENTS COUNT : \t $streamingEvents")
+      totalEvents += streamingEvents
+      log.info(s" TOTAL EVENTS : \t $totalEvents")
+      val streamingRegisters = rdd.collect()
+      if (!rdd.isEmpty())
+        streamingRegisters.foreach(row => assert(data1.contains(row)))
+    })
+    ssc.start()
+    ssc.awaitTerminationOrTimeout(timeoutStreaming)
+    ssc.stop()
+
+    assert(totalEvents.value === 2)
+  }
+}