+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.0 (upcoming)
 
 * Join transformation
+* Drop duplicates transformation
 
 ## 2.0.0 (April 03, 2018)